ActiveQuestsStep: read todos via QuestEventHandler, export full row id
Two corrections after testing the first version against a live game state:

1. todo_checked was reading bit i of QuestWork.Variables[5], which is
   never the right place. Variables is per-quest scratch space, not a
   normalized objective bitmap. Use QuestEventHandler.IsTodoChecked
   (resolved via EventFramework.GetEventHandlerById(questId)) — the
   same path the FFXIV ToDoList addon uses internally. Sequence 255
   (turn-in stage) is short-circuited to all-false since no objectives
   are tracked at that point.

2. id was exported as the raw QuestWork.QuestId, the lower 16 bits of
   the quest's row id. The Excel `Quest` sheet uses the full 32-bit
   id (offset by 0x10000), so consumers had to do the math themselves.
   Switch the field to uint and add 0x10000 at write time.

+++ Auracite/Package.cs
@@ -80,7 +80,7 @@ public class Appearance
 [SuppressMessage("ReSharper", "InconsistentNaming")]
 public class ActiveQuest
 {
-    public ushort id;
+    public uint id;
     public byte sequence;
     public byte flags;
     public byte accept_class_job;
